mainApp/serializers.py
# your_app/serializers.py

from django.contrib.auth.models import User
from rest_framework import serializers
from .models import Item

class UserRegistrationSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True, required=True, style={'input_type': 'password'})
    password2 = serializers.CharField(write_only=True, required=True, label="Confirm Password", style={'input_type': 'password'})

    class Meta:
        model = User
        fields = ('username', 'email', 'password', 'password2')

    def validate(self, attrs):
        if attrs['password'] != attrs['password2']:
            raise serializers.ValidationError({"password": "Passwords must match."})
        return attrs

    def create(self, validated_data):
        user = User.objects.create_user(
            username=validated_data['username'],
            email=validated_data.get('email'),
            password=validated_data['password']
        )
        return user

class UserExistenceSerializer(serializers.Serializer):
    username = serializers.CharField(required=True)

class ItemSerializer(serializers.ModelSerializer):
    owner = serializers.ReadOnlyField(source='owner.username')  # Display owner's username

    class Meta:
        model = Item
        fields = ['id', 'name', 'description', 'owner', 'created_at', 'updated_at']
